feat(quiz): refactor quiz components and hooks for answer quiz functionality
