+++ app/(tabs)/questions/index.tsx
@@ -2,10 +2,10 @@ import { PrimaryButton } from '@/components/ui/button/PrimaryButton'
 import { Colors } from '@/constants/Colors'
 import { useGlobalContext } from '@/context/GlobalContext'
 import { Tables } from '@/database.types'
-import { GroupButton } from '@/features/solve-problems/components/GroupButton'
-import { GroupSelectionHeader } from '@/features/solve-problems/components/GroupSelectionHeader'
-import { useNextQuiz } from '@/features/solve-problems/hooks/useNextQuiz'
-import { useSyncUnansweredQuizIds } from '@/features/solve-problems/hooks/useSyncUnansweredQuizIds'
+import { GroupButton } from '@/features/answer-quiz/components/GroupButton'
+import { GroupSelectionHeader } from '@/features/answer-quiz/components/GroupSelectionHeader'
+import { useNextQuiz } from '@/features/answer-quiz/hooks/useNextQuiz'
+import { useSyncUnansweredQuizIds } from '@/features/answer-quiz/hooks/useSyncUnansweredQuizIds'
 import { supabase } from '@/utils/supabase'
 import { showErrorToast } from '@/utils/toast'
 import { useQuery } from '@tanstack/react-query'
