+++ features/answer-quiz/components/GroupSelectionHeader.tsx
